Remove context.py and move global variables to hoomd namespace
